requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.2
fonttools==4.40.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
h5py==3.9.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.25.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
medutils-mri==0.1.1
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
numpy==1.25.0
opencv-python==4.8.1.78
outcome==1.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.1.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.14.0
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==68.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchgeometry==0.1.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.12
wcwidth==0.2.6
wheel==0.40.0
wsproto==1.2.0
zipp==3.15.0
